requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
structlog>=24.0.0
fhir.resources>=7.0.0
hl7apy>=1.3.0
httpx>=0.27.0
tenacity>=8.2.0
xmltodict>=0.13.0
